pyproject.toml
# pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "syslog-generator"
version = "1.0.0"
description = "A tool for generating realistic syslog messages for testing"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Logging",
    "Topic :: Software Development :: Testing",
]
dependencies = [
    "pyyaml>=6.0",
    "faker>=18.0.0",
    "colorama>=0.4.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "pylint>=2.17.0",
]

[project.scripts]
syslog-generator = "syslog_generator.main:main"

[project.urls]
Homepage = "https://github.com/yourusername/syslog-generator"
Repository = "https://github.com/yourusername/syslog-generator"

[tool.setuptools.packages.find]
where = ["."]
include = ["syslog_generator*"]

[tool.setuptools.package-data]
syslog_generator = ["*.yaml", "*.json"]
